various bugfixes

+++ pairinteraction_gui/pairinteraction/app.py
@@ -450,7 +450,8 @@ class MainWindow(QtWidgets.QMainWindow):
         self.ui = Ui_plotwindow()
         self.ui.setupUi(self)
 
-        self.invalidQuantumnumbers = [False, False, False, False]
+        self.invalidQuantumnumbers = {}
+        self.invalidQuantumnumbersMessage = ""
 
         self.samebasis = False
 
@@ -569,26 +570,19 @@ class MainWindow(QtWidgets.QMainWindow):
         conn.close()
 
         # TODO !!!!!! numBlocks kann auch hoeher als 3 sein!
-        self.buffer_basis = [{}, {}, {}]
-        self.buffer_energies = [{}, {}, {}]
-        self.buffer_positions = [{}, {}, {}]
-        self.buffer_boolarr = [{}, {}, {}]
-        self.buffer_basis_potential = {}
-        self.buffer_energies_potential = {}
-        self.buffer_positions_potential = {}
+        self.buffer_basis = {}
+        self.buffer_energies = {}
+        self.buffer_positions = {}
+        self.buffer_boolarr = {}
 
         self.buffer_energiesMap = [{}, {}, {}]
         self.buffer_positionsMap = [{}, {}, {}]
         self.buffer_overlapMap = [{}, {}, {}]
-        self.buffer_energiesMap_potential = {}
-        self.buffer_positionsMap_potential = {}
-        self.buffer_overlapMap_potential = {}
 
         self.lines_buffer_minIdx = {}
-        self.colormap_buffer_minIdx_potential = 0
         self.colormap_buffer_minIdx_field = [0] * 3
-        self.lines_buffer_minIdx_field = [0] * 3
-        self.iSelected = {}
+        self.lines_buffer_minIdx_field = {}
+        """self.iSelected = {}"""
 
         """self.ui.colorbutton_plot_nosym.setColor(self.symmetrycolors[0])
         self.ui.colorbutton_plot_sym.setColor(self.symmetrycolors[1])
@@ -696,22 +690,12 @@ class MainWindow(QtWidgets.QMainWindow):
         # Connect signals and slots
         self.thread.criticalsignal.connect(self.showCriticalMessage)
 
-        self.ui.spinbox_system_n1.valueChanged.connect(self.validateQuantumnumbers)
-        self.ui.spinbox_system_n2.valueChanged.connect(self.validateQuantumnumbers)
-        self.ui.spinbox_system_l1.valueChanged.connect(self.validateQuantumnumbers)
-        self.ui.spinbox_system_l2.valueChanged.connect(self.validateQuantumnumbers)
-        self.ui.spinbox_system_j1.valueChanged.connect(self.validateQuantumnumbers)
-        self.ui.spinbox_system_j2.valueChanged.connect(self.validateQuantumnumbers)
-        self.ui.spinbox_system_m1.valueChanged.connect(self.validateQuantumnumbers)
-        self.ui.spinbox_system_m2.valueChanged.connect(self.validateQuantumnumbers)
-        self.ui.spinbox_plot_n1.valueChanged.connect(self.validateQuantumnumbers)
-        self.ui.spinbox_plot_n2.valueChanged.connect(self.validateQuantumnumbers)
-        self.ui.spinbox_plot_l1.valueChanged.connect(self.validateQuantumnumbers)
-        self.ui.spinbox_plot_l2.valueChanged.connect(self.validateQuantumnumbers)
-        self.ui.spinbox_plot_j1.valueChanged.connect(self.validateQuantumnumbers)
-        self.ui.spinbox_plot_j2.valueChanged.connect(self.validateQuantumnumbers)
-        self.ui.spinbox_plot_m1.valueChanged.connect(self.validateQuantumnumbers)
-        self.ui.spinbox_plot_m2.valueChanged.connect(self.validateQuantumnumbers)
+        for system in ["system", "plot"]:
+            for number in [1, 2]:
+                for q in "nljm":
+                    getattr(self.ui, f"spinbox_{system}_{q}{number}").valueChanged.connect(
+                        self.validateOneQuantumnumbers
+                    )
 
         self.ui.spinbox_system_j1.editingFinished.connect(self.validateSpinLike)
         self.ui.spinbox_system_j2.editingFinished.connect(self.validateSpinLike)
@@ -728,8 +712,8 @@ class MainWindow(QtWidgets.QMainWindow):
         self.ui.combobox_system_species2.currentIndexChanged[str].connect(self.forbidSamebasis)
         self.ui.combobox_system_species1.currentIndexChanged[str].connect(self.defaultQuantumnumbers)
         self.ui.combobox_system_species2.currentIndexChanged[str].connect(self.defaultQuantumnumbers)
-        self.ui.combobox_system_species1.currentIndexChanged.connect(self.validateQuantumnumbers)
-        self.ui.combobox_system_species2.currentIndexChanged.connect(self.validateQuantumnumbers)
+        self.ui.combobox_system_species1.currentIndexChanged.connect(self.validateAllQuantumnumbers)
+        self.ui.combobox_system_species2.currentIndexChanged.connect(self.validateAllQuantumnumbers)
 
         self.ui.radiobutton_system_pairbasisDefined.toggled.connect(self.togglePairbasis)
         self.ui.radiobutton_plot_overlapDefined.toggled.connect(self.toggleOverlapstate)
@@ -1501,11 +1485,17 @@ class MainWindow(QtWidgets.QMainWindow):
                 os.remove(basisfile)
 
                 # clear variables
-                self.lines_buffer_minIdx_field = {}
-                self.buffer_basis = {}
-                self.buffer_energies = {}
-                self.buffer_positions = {}
-                self.buffer_boolarr = {}
+                for buffer in [
+                    self.lines_buffer_minIdx_field,
+                    self.buffer_basis,
+                    self.buffer_energies,
+                    self.buffer_positions,
+                    self.buffer_boolarr,
+                ]:
+                    _bns = list(buffer.keys())
+                    for _bn in _bns:
+                        if _bn == bn or bn == NO_BN:
+                            del buffer[_bn]
 
                 if len(self.storage_states[idx]) == 1:
                     self.labelprob = None
@@ -1544,7 +1534,7 @@ class MainWindow(QtWidgets.QMainWindow):
                     # save data
                     self.storage_data[idx].append([filestep, blocknumber, filename])
 
-                    if "pkl" in filename:
+                    if filename.endswith(".pkl"):
                         with open(filename, "rb") as f:
                             pkl_data = pickle.load(f)
                         energies, basis, params = pkl_data["energies"], pkl_data["basis"], pkl_data["params"]
@@ -2102,7 +2092,8 @@ class MainWindow(QtWidgets.QMainWindow):
 
                         while (
                             self.lines_buffer_minIdx_field[blocknumber] in self.buffer_basis[blocknumber].keys()
-                            and self.lines_buffer_minIdx_field[blocknumber] + 1 in self.buffer_basis[blocknumber].keys()
+                            and (self.lines_buffer_minIdx_field[blocknumber] + 1)
+                            in self.buffer_basis[blocknumber].keys()
                         ):
                             # determine the data to plot
                             overlap = np.abs(
@@ -2274,25 +2265,18 @@ class MainWindow(QtWidgets.QMainWindow):
             and self.thread.dataqueue_potential.empty()
         ):
             # Delete buffers
-            self.buffer_basis = [{}, {}, {}]
-            self.buffer_energies = [{}, {}, {}]
-            self.buffer_positions = [{}, {}, {}]
-            self.buffer_boolarr = [{}, {}, {}]
-            self.buffer_basis_potential = {}
-            self.buffer_energies_potential = {}
-            self.buffer_positions_potential = {}
+            self.buffer_basis = {}
+            self.buffer_energies = {}
+            self.buffer_positions = {}
+            self.buffer_boolarr = {}
 
             self.buffer_energiesMap = [{}, {}, {}]
             self.buffer_positionsMap = [{}, {}, {}]
             self.buffer_overlapMap = [{}, {}, {}]
-            self.buffer_energiesMap_potential = {}
-            self.buffer_positionsMap_potential = {}
-            self.buffer_overlapMap_potential = {}
 
             self.lines_buffer_minIdx = {}
-            self.colormap_buffer_minIdx_potential = 0
             self.colormap_buffer_minIdx_field = [0] * 3
-            self.lines_buffer_minIdx_field = [0] * 3
+            self.lines_buffer_minIdx_field = {}
 
             # Delete c++ process
             if self.proc is not None:
@@ -2513,6 +2497,7 @@ class MainWindow(QtWidgets.QMainWindow):
     def toggleOverlapstate(self):
         checked = self.ui.radiobutton_plot_overlapDefined.isChecked()
         self.ui.widget_plot_qn.setEnabled(checked)
+        self.validateAllQuantumnumbers()
 
     @QtCore.pyqtSlot(bool)  # TODO
     def toggleSymmetrization(self):
@@ -2640,12 +2625,20 @@ class MainWindow(QtWidgets.QMainWindow):
         if sender.value() != -1 and sender.value() < 2:
             sender.setValue(2)
 
-    @QtCore.pyqtSlot()
-    def validateQuantumnumbers(self):
-        sender_name = self.sender().objectName()
-        system = "plot" if "plot" in sender_name else "system"
-        number = 2 if "2" in sender_name else 1
-        i = number - 1 + (2 if system == "plot" else 0)
+    def validateAllQuantumnumbers(self):
+        self.invalidQuantumnumbers = {}
+        check_systems = ["system", "plot"] if self.ui.radiobutton_plot_overlapDefined.isChecked() else ["system"]
+        for system in check_systems:
+            for number in [1, 2]:
+                self.validateOneQuantumnumbers(system, number)
+
+    def validateOneQuantumnumbers(self, system=None, number=None):
+        if system is None or number is None:
+            sender_name = self.sender().objectName()
+            system = "plot" if "plot" in sender_name else "system"
+            number = 2 if "2" in sender_name else 1
+        assert system in ["system", "plot"] and number in [1, 2]
+        name = system + str(number)
 
         qns = []
         for q in "nljm":
@@ -2659,18 +2652,19 @@ class MainWindow(QtWidgets.QMainWindow):
         is_plot = system == "plot"
         if l >= n and not (is_plot and PLOT_ALL in [n, l]):
             err = True
-        if abs(m) > j and not (is_plot and PLOT_ALL in [j, m]):
+        elif abs(m) > j and not (is_plot and PLOT_ALL in [j, m]):
             err = True
-        if abs(l - j) > s and not (is_plot and PLOT_ALL in [l, j]):
+        elif abs(l - j) > s and not (is_plot and PLOT_ALL in [l, j]):
             err = True
-        if (j - s) % 1 != 0 and not (is_plot and PLOT_ALL in [j]):
+        elif (j - s) % 1 != 0 and not (is_plot and PLOT_ALL in [j]):
             err = True
+        self.invalidQuantumnumbers[name] = err
 
-        self.invalidQuantumnumbers[i] = err
-        if np.any(self.invalidQuantumnumbers):
-            self.ui.statusbar.showMessage("Invalide quantum numbers specified.")
-        else:
-            self.ui.statusbar.showMessage("")
+        msg = ", ".join([k for k, v in self.invalidQuantumnumbers.items() if v])
+        if msg != "":
+            msg = f"Invalid quantum numbers specified ({msg})."
+        self.ui.statusbar.showMessage(msg)
+        self.invalidQuantumnumbersMessage = msg
 
     def _validateQnGeneral(self, _type="integer", orPlotAll=False):
         assert _type in ["integer", "integerpositive", "spinlike", "spinlikepositive"]
@@ -2718,8 +2712,8 @@ class MainWindow(QtWidgets.QMainWindow):
             if focused_widget is not None:
                 focused_widget.clearFocus()
 
-            if np.any(self.invalidQuantumnumbers):
-                QtWidgets.QMessageBox.critical(self, "Message", "Invalide quantum numbers specified.")
+            if any(self.invalidQuantumnumbers.values()):
+                QtWidgets.QMessageBox.critical(self, "Message", self.invalidQuantumnumbersMessage)
 
             elif (
                 self.ui.radiobutton_system_missingWhittaker.isChecked()
@@ -3078,7 +3072,7 @@ class MainWindow(QtWidgets.QMainWindow):
                 if self.ui.checkbox_use_python_api.isChecked():
                     path_python = os.path.join(self.path_base, "start_pipy.py")
                     self.proc = subprocess.Popen(
-                        [path_python, "--run_gui", "-c", self.path_conf, "-o", self.path_cache],
+                        [sys.executable, path_python, "--run_gui", "-c", self.path_conf, "-o", self.path_cache],
                         stdout=subprocess.PIPE,
                         stderr=subprocess.STDOUT,
                         cwd=self.path_workingdir,
@@ -3147,22 +3141,23 @@ class MainWindow(QtWidgets.QMainWindow):
 
         try:
             # save plot
-            plotitem = [
-                self.ui.graphicsview_field1_plot,
-                self.ui.graphicsview_field2_plot,
-                self.ui.graphicsview_potential_plot,
-            ][idx].getPlotItem()
-            exporter = exporters.ImageExporter(plotitem)
-            exporter.parameters()["width"] = 2000
-            exporter.parameters()["height"] = 2000
-            exporter.parameters()["antialias"] = True
-            image = exporter.export(toBytes=True)
+            if getattr(self, "savePlot", True):
+                plotitem = [
+                    self.ui.graphicsview_field1_plot,
+                    self.ui.graphicsview_field2_plot,
+                    self.ui.graphicsview_potential_plot,
+                ][idx].getPlotItem()
+                exporter = exporters.ImageExporter(plotitem)
+                exporter.parameters()["width"] = 2000
+                exporter.parameters()["height"] = 2000
+                exporter.parameters()["antialias"] = True
+                image = exporter.export(toBytes=True)
 
-            buffer = QtCore.QBuffer()
-            buffer.open(QtCore.QIODevice.WriteOnly)
-            image = image.scaled(1500, 1500, QtCore.Qt.KeepAspectRatio, QtCore.Qt.SmoothTransformation)
-            image.save(buffer, "PNG")
-            ziparchive.writestr("plot.png", buffer.data())
+                buffer = QtCore.QBuffer()
+                buffer.open(QtCore.QIODevice.WriteOnly)
+                image = image.scaled(1500, 1500, QtCore.Qt.KeepAspectRatio, QtCore.Qt.SmoothTransformation)
+                image.save(buffer, "PNG")
+                ziparchive.writestr("plot.png", buffer.data())
 
             # save configuration
             ziparchive.writestr("settings.sconf", self.storage_configuration[idx][0])
@@ -3213,12 +3208,10 @@ class MainWindow(QtWidgets.QMainWindow):
                 data["numOverlapvectors"] = self.stateidx_field[idx][NO_BN].shape[0]
                 data["overlapvectors"] = self.stateidx_field[idx][NO_BN]
             elif len(self.stateidx_field[idx]) > 0:
-                data["numOverlapvectors"] = [
-                    [self.stateidx_field[idx][k].shape[0]] for k in sorted(self.stateidx_field[idx].keys())
-                ]
-                data["overlapvectors"] = [
-                    [self.stateidx_field[idx][k]] for k in sorted(self.stateidx_field[idx].keys())
-                ]
+                _stateidx = self.stateidx_field[idx]
+                keys = [k for k in _stateidx.keys() if _stateidx[k] is not None]
+                data["numOverlapvectors"] = [[_stateidx[k].shape[0]] for k in keys]
+                data["overlapvectors"] = [[_stateidx[k]] for k in keys]
 
             # save data
             # TODO Variablen an anderer Stelle anlegen
@@ -3240,7 +3233,7 @@ class MainWindow(QtWidgets.QMainWindow):
             filestep_last = None
 
             for filestep, _blocknumber, filename in sorted(self.storage_data[idx], key=itemgetter(0, 1)):
-                if "pkl" in filename:
+                if filename.endswith(".pkl"):
                     with open(filename, "rb") as f:
                         pkl_data = pickle.load(f)
                     energies, basis, params = pkl_data["energies"], pkl_data["basis"], pkl_data["params"]
@@ -3506,21 +3499,14 @@ class MainWindow(QtWidgets.QMainWindow):
 
     @QtCore.pyqtSlot()
     def clearCache(self):
-        files = [
-            "cache_elements.db",
-            "cache_matrix_complex.db",
-            "cache_matrix_real.db",
-            "cache_matrix_complex",
-            "cache_matrix_real",
-            "cache_matrix_complex_new",
-            "cache_matrix_real_new",
-        ]  # TODO: sicherstellen, dass gleiche Namen wie im C++ Programm
-        for file in files:
-            path = os.path.join(self.path_cache, file)
+        for name in os.listdir(self.path_cache):
+            path = os.path.join(self.path_cache, name)
             if os.path.isfile(path):
-                os.remove(path)
+                if (name.startswith("cache_elements") or name.startswith("cache_matrix")) and name.endswith(".db"):
+                    os.remove(path)
             elif os.path.isdir(path):
-                shutil.rmtree(path)
+                if name.startswith("cache_matrix"):
+                    shutil.rmtree(path)
 
         if os.path.isdir(self.path_cache_wignerd):
             shutil.rmtree(self.path_cache_wignerd)
